Issue #340 - extract Stack as a proper class.
The users have to use ProgressBar#getStackId() explicitly now
